test(openai-compatible): preserve providers in Nebius example fixtures
